+++ frontend/src/components-business/form-editor/RelatedAnnotationEditor.vue
@@ -25,7 +25,7 @@ const embeddedSwitch = ref(false)
 const searchProps = {
     autoFocus: true,
     query: (client: HttpClient) => (offset: number, limit: number, search: string) => client.annotation.list({offset, limit, query: search, type: props.metaType, target: props.target}),
-    historyList: (client: HttpClient) => client.searchUtil.history.annotations,
+    historyList: (client: HttpClient) => (_: number) => client.searchUtil.history.annotations(props.metaType),
     historyPush: (client: HttpClient) => (item: Annotation) => client.searchUtil.history.push({type: "ANNOTATION", id: item.id}),
     mapOption: (item: Annotation) => ({label: item.name, value: `${item.id}`}),
     onPick: (item: Annotation) => {