+++ frontend/src/components-business/top-bar/QuerySchema/QueryResultCompileError.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { Block } from "@/components/universal"
 import { CompileError } from "@/functions/http-client/api/util-query"
+import { queryCompileErrorTranslate } from "@/constants/translate"
 
 defineProps<{
     e: CompileError
@@ -11,6 +12,6 @@ defineProps<{
 
 <template>
     <Block class="p-1 mt-1 is-font-size-small" :color="type">
-        [{{e.code}}]<b v-if="e.happenPosition">({{e.happenPosition.begin}}, {{e.happenPosition.end ?? e.happenPosition.begin}})</b>: {{e.message}}
+        [{{e.code}}]<b v-if="e.happenPosition">({{e.happenPosition.begin}}, {{e.happenPosition.end ?? e.happenPosition.begin}})</b>: {{queryCompileErrorTranslate(e)}}
     </Block>
 </template>