+++ frontend/src/components-business/top-bar/QuerySchema/QueryResultPlan.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { QueryPlan } from "@/functions/http-client/api/util-query"
-import QueryResultPlanOrderList from "./QueryResultPlanOrderList.vue"
+import QueryResultPlanSortList from "./QueryResultPlanSortList.vue"
 import QueryResultPlanFilterItem from "./QueryResultPlanFilter.vue"
 import QueryResultPlanElement from "./QueryResultPlanElement.vue"
 
@@ -14,6 +14,6 @@ defineProps<{
     <div class="px-1 pb-1">
         <QueryResultPlanElement v-for="element in plan.elements" class="mt-1" :element-group="element"/>
         <QueryResultPlanFilterItem v-for="filterGroup in plan.filters" class="mt-1" :filter-group="filterGroup"/>
-        <QueryResultPlanOrderList v-if="plan.orders.length > 0" class="mt-1" :orders="plan.orders"/>
+        <QueryResultPlanSortList v-if="plan.sorts.length > 0" class="mt-1" :sorts="plan.sorts"/>
     </div>
 </template>