+++ frontend/src/components-business/top-bar/QuerySchema/QueryResultPlanSortList.vue
@@ -3,7 +3,7 @@ import { Flex } from "@/components/layout"
 import { QUERY_FIELD_NAMES } from "@/constants/translate"
 
 defineProps<{
-    orders: string[]
+    sorts: string[]
 }>()
 
 </script>
@@ -12,7 +12,7 @@ defineProps<{
     <Flex>
         <div :class="$style.name"><b>排序</b></div>
         <div :class="$style.list">
-            <span v-for="(o, idx) in orders">
+            <span v-for="(o, idx) in sorts">
                 <b v-if="idx > 0" class="ml-1 mr-2">,</b>
                 <b>{{o.charAt(0) === '-' ? '-' : '+'}}</b>{{QUERY_FIELD_NAMES[o.substring(1)] ?? o.substring(1)}}
             </span>