+++ frontend/src/views/Main/Partition/Partition.vue
@@ -16,7 +16,7 @@ const {
 <template>
     <BrowserTeleport to="top-bar">
         <CollectionModeButton class="mr-1" v-model:value="collectionMode"/>
-        <SearchBox placeholder="在此处搜索" v-model:value="querySchema.queryInputText.value" :enable-drop-button="!!querySchema.query.value" v-model:active-drop-button="querySchema.expanded.value" :schema="querySchema.schema.value"/>
+        <SearchBox placeholder="在此处搜索" dialect="ILLUST" v-model:value="querySchema.queryInputText.value" :schema="querySchema.schema.value"/>
         <Separator/>
         <div class="mr-1 ml-1 flex-item align-center no-grow-shrink">
             <Tag>{{ total.count }}项</Tag>